Converted login.js to typescript
